fix(iRestHttp): Use encodeURIComponent on XMLSERVICE params
Using encodeURI on the entire XMLSERVICE call URI will not encode
non-url-safe characters such as # and & properly when they appear inside
a parameter value (or by extension, anywhere in the XML document). This
causes an error in XMLSERVICE if the input XML document contains # or &.
Instead, calling encodeURIComponent individually on each parameter in
the XMLSERVICE call URI will safely encode all possible characters.